flytte apikall ut
